requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
attrs==20.3.0
    # via jsonschema
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
idna==2.10
    # via requests
jsonschema==3.2.0
    # via valohai-yaml
py4j==0.10.9
    # via pyspark
pyrsistent==0.17.3
    # via jsonschema
pyspark-stubs==3.0.0.post1
    # via -r requirements.in
pyspark==3.0.2
    # via
    #   -r requirements.in
    #   pyspark-stubs
pyyaml==5.4.1
    # via valohai-yaml
requests==2.25.1
    # via valohai-utils
six==1.15.0
    # via jsonschema
tqdm==4.58.0
    # via valohai-utils
urllib3==1.26.3
    # via requests
valohai-utils==0.1.2
    # via -r requirements.in
valohai-yaml==0.13.0
    # via valohai-utils

# The following packages are considered to be unsafe in a requirements file:
# setuptools
